Remove duplicated information
Users are confused by seeing a subset of the fields in the standard 
organogram template in the "What to include section".

This makes them unsure if all the fields in the template are actually 
required.

The standard organogram template should be the single source of truth 
for infromation that needs to be published.

Therefore remove the bulleted lists and refer to the organogram as the 
single source of truth.