Move details about schemas to the appropriate step
We've seen in the lab that users ignore this field because they're 
confused about the manu options and they're not sure what to choose. The 
guidance on what schema to choose comes too late, after the user will 
have clicked Publish.

data.gov.uk have also received support queries because users haven't 
selected a schema, so this change also addresses the issue in the 
support tickets.

Also changed formatting from inline code to bold because the text is 
user interface text. Plus, users, who are not developers, were confused 
by the "odd formatting" of inline code.