Use "create organogram dataset" and "add organogram to dataset"
Do not use "update an organogram". That would mean modifying data that 
was uploaded, not adding a new organogram to a dataset.

Users pointed out:
"I wouldn’t be *updating* an organogram, I would be *adding* an 
organogram to an organogram dataset."
"I add the new CSV data to the existing dataset which was created 
automatically by the previous upload process."

Added text to explain when it's appropriate to use each set of 
instructions to avoid users creating a new dataset every time they 
upload an organogram.

Added a bit more detail to the instructions of how to add an organogram 
to a dataset.