Clarify permissions and preferred registration approach
The original docs state that admins and editors can invite people to the organisation,
but in reality only admins can do this. I've updated the text to clarify this.

In addition, the docs don't particularly prescribe one way or the other whether or not
you should be added by your admin or by GOV.UK 2nd Line. As GOV.UK 2nd Line direct
you back to your organisation admin as a first port of call, we should make it clearer in this
document that that is the step we advise people to make first, otherwise they may lose
a few days waiting for a response from 2nd line only to be told the same thing.