Update link to Gemini 2.3 Schematron
While in this documentation we refer to Gemini 2.3 Schematron, we are actually
linking to Gemini 2.1 Schematron. This was causing confusion for users who were
expecting to find the 2.3 Schematron.

This updates the link to Gemini 2.3 Schematron and points users to the file
which is available in the alphagov/ckanext-spatial repo - it is not available
in the datagovuk/ckanext-spatial repo.  Since we seem to be using the alphagov
repo and not the datagovuk repo for ckanext-spatial we update the remaining
links on this page to point to the alpahgov repo for consistency.